Added exit from fullscreen mode, and tracking the fullscreen state
