+++ src/screens/InspectionCapture/settings.js
@@ -1,35 +1,12 @@
-import React, { forwardRef, useImperativeHandle, useState } from 'react';
-import { Pressable, useWindowDimensions, View } from 'react-native';
+import React, { forwardRef, useCallback, useImperativeHandle, useMemo, useState } from 'react';
+import { Platform, Pressable, useWindowDimensions, View } from 'react-native';
 import { useTheme, Menu, List } from 'react-native-paper';
 import PropTypes from 'prop-types';
 
 import { Actions } from '@monkvision/camera';
 
 import styles from './styles';
-
-const requestScreenful = () => {
-  document.fullscreenEnabled = document.fullscreenEnabled
-   || document.mozFullScreenEnabled || document.documentElement.webkitRequestFullScreen;
-
-  function requestFullscreen(element) {
-    if (element.requestFullscreen) {
-      element.requestFullscreen();
-    } else if (element.mozRequestFullScreen) {
-      element.mozRequestFullScreen();
-    } else if (element.webkitRequestFullScreen) {
-      element.webkitRequestFullScreen(Element.ALLOW_KEYBOARD_INPUT);
-    }
-  }
-
-  if (document.fullscreenEnabled) {
-    requestFullscreen(document.documentElement);
-  }
-};
-
-const modals = {
-  default: [{ title: 'Resolution', value: 'resolution' }, { title: 'Fullscreen', value: 'fullscreen' }],
-  resolution: [{ title: 'FHD', value: 'FHD' }, { title: 'QHD', value: 'QHD' }],
-};
+import useFullscreen from './useFullscreen';
 
 const Settings = forwardRef(({ settings }, ref) => {
   const { colors } = useTheme();
@@ -37,33 +14,39 @@ const Settings = forwardRef(({ settings }, ref) => {
   const [modal, setModal] = useState({ visible: false, name: null });
   const handleClose = () => setModal({ visible: false, name: null });
 
-  const handleSelect = (name) => {
+  const { isFullscreen, requestFullscreen } = useFullscreen();
+
+  const modals = useMemo(() => ({
+    default: [{ title: 'Resolution', value: 'resolution' }, { hidden: Platform.OS !== 'web', title: isFullscreen ? 'Exit fullscreen' : 'Fullscreen', value: 'fullscreen', selected: isFullscreen }],
+    resolution: [{ title: 'FHD', value: 'FHD' }, { title: 'QHD', value: 'QHD' }],
+  }), [isFullscreen]);
+
+  const handleSelect = useCallback((name) => {
     // select only value that are not one of the `modals` keys
-    if (name === 'fullscreen') { requestScreenful(); return; }
+    if (name === 'fullscreen') { requestFullscreen(); return; }
     if (Object.keys(modals).includes(name)) { setModal({ visible: true, name }); return; }
 
     settings.dispatch({
       type: Actions.settings.UPDATE_SETTINGS,
       payload: { [modal.name]: name },
     });
-  };
+  }, [modals, requestFullscreen, modal.name]);
 
-  useImperativeHandle(ref, () => ({
-    open: () => handleSelect('default'),
-  }));
+  useImperativeHandle(ref, () => ({ open: () => handleSelect('default') }));
 
   if (!modal.visible) { return null; }
   return (
     <>
       <View style={[styles.settings, { backgroundColor: colors.background }]}>
         <List.Subheader>Settings</List.Subheader>
-        {modals[modal.name].map((item) => (
+        {modals[modal.name].map((item) => !item.hidden && (
           <Menu.Item
             onPress={() => handleSelect(item.value)}
             title={item.title}
             key={item.value}
             style={[styles.settingItem, { backgroundColor: settings.state[modal.name] === item.value
-              ? colors.highlightBoneColor
+               || item.selected
+              ? colors.surface
               : colors.background,
             }]}
           />
